Fix asyncWriter architecture: eliminate bypassed logging setup
main.go's asyncWriter was immediately overridden by cmd/root.go's
configureLogging(), leaving the async buffer unused and leaking a file
handle. Move asyncWriter to the cmd package and wrap all log destinations
(journal, file, stderr) in configureLogging() so there is a single
source of truth for log configuration with proper async buffering.
